feat: split spiking_results from reportings
