+++ cerebellar_models/analysis/__init__.py
@@ -11,7 +11,7 @@ from cerebellar_models.analysis.report import (
     PlotTypeInfo,
     Report,
 )
-from cerebellar_models.analysis.spiking_results import (
+from cerebellar_models.analysis.spike_plots import (
     BasicSimulationReport,
     FiringRatesPlot,
     FrequencyPlot,
@@ -20,6 +20,13 @@ from cerebellar_models.analysis.spiking_results import (
     SimResultsTable,
     SpikePlot,
     SpikeSimulationReport,
+)
+from cerebellar_models.analysis.spiking_results import (
+    SpikingResults,
+    extract_isis,
+    get_correlation_coefficients,
+    get_firing_rates,
+    get_frequencies,
 )
 from cerebellar_models.analysis.structure_analysis import (
     CellPlacement3D,
@@ -45,6 +52,7 @@ __all__ = [
     "FrequencyPlot",
     "SpikePlot",
     "SpikeSimulationReport",
+    "SpikingResults",
     "SimResultsTable",
     "ISIPlot",
     "RasterPSTHPlot",
@@ -54,4 +62,8 @@ __all__ = [
     "CellPlacement3D",
     "StructureReport",
     "RunStructureReport",
+    "extract_isis",
+    "get_correlation_coefficients",
+    "get_firing_rates",
+    "get_frequencies",
 ]
